requirements.txt
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.0
attrs==25.3.0
bleach==6.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
emoji==2.14.1
et_xmlfile==2.0.0
executing==2.1.0
fastapi==0.124.0
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.3.2
h11==0.16.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kaggle==1.7.4.5
kiwisolver==1.4.8
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mplfinance==0.12.10b0
narwhals==1.48.1
nest-asyncio==1.6.0
numpy==2.4.4
openpyxl==3.1.5
packaging==24.2
pandas==3.0.2
pandas_ta==0.3.14b0
parso==0.8.4
pillow==11.3.0
platformdirs==4.3.6
plotly==6.2.0
prompt_toolkit==3.0.50
protobuf==6.31.1
psutil==6.1.1
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.1
PyMySQL==1.1.2
pyodbc==5.3.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-pptx==1.0.2
python-slugify==8.0.4
pytz==2025.2
pywin32==308
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.49
stack-data==0.6.3
starlette==0.50.0
streamlit==1.47.1
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.5.0
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
xgboost==3.1.2
xlrd==2.0.2
xlsxwriter==3.2.9
